+++ backend/auth_service/.gitignore
