+++ backend/auth_service/api/controllers/emailController.js
@@ -6,15 +6,13 @@ export const sendEmail = async (request, reply) => {
     console.log("Sending email... ", request.body);
     const { to } = request.body;
 
-    // Generar OTP
     const otp = crypto.randomInt(100000, 999999).toString();
 
-    // Convertir OTP en JWT
     const otpToken = generateOTPToken(otp);
 
     const msg = {
         to,
-        from: "transcendencetheboys@gmail.com",
+        from: process.env.THE_BOYS,
         subject: "Your OTP Code",
         text: `Your OTP code is ${otp}. Use this code to login to your account. It expires in 5 minutes.`,
     };
@@ -41,14 +39,3 @@ export const validateOTP = async (request, reply) => {
         reply.send({ success: false, message: "Invalid or expired OTP!" });
     }
 };
-/*export const validateOTP = async (request, reply) => {
-    const { otp, secret } = request.body; // El usuario debe enviar su OTP y su secret almacenado
-
-    const isValid = verifyOTP(otp, secret);
-
-    if (isValid) {
-        reply.send({ success: true, message: "Código válido, autenticación completada!" });
-    } else {
-        reply.send({ success: false, message: "Código inválido o expirado!" });
-    }
-};*/
