+++ backend/auth_service/api/routes/authRoutes.js
