+++ backend/auth_service/api/routes/emailRoutes.js
