+++ env_example
@@ -7,10 +7,14 @@ NODE_ENV="production"
 GAME_DATABASE_URL="file:../data/game.db"
 
 # googleAuth variables
-GOOGLE_CLIENT_ID=626578837925-d6ql0k67ih2bo53oekrmm612keue64f4.apps.googleusercontent.com
-GOOGLE_CLIENT_SECRET=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
+GOOGLE_CLIENT_ID=341612200177-d9uls69p5itl1in6b0i9di5ogj750478.apps.googleusercontent.com
+GOOGLE_CLIENT_SECRET=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
 GOOGLE_CALLBACK_URL=https://localhost:4433/auth/google/callback
 FASTIFY_SECURE_SECRET=96e86052a27b52b634bba5415dcf80909d4c0d14a7e44826ccb4762ac4f4b656
+SENDGRID_API_KEY=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
+REDIRECT_URL=https://localhost:4433
+THE_BOYS=transcendencetheboys@gmail.com
+NODE_ENV="production"
 
 # Password for the 'elastic' user (at least 6 characters)
 ELASTIC_USERNAME=elastic
