+++ package.json
@@ -4,7 +4,7 @@
 	"type": "module",
 	"private": true,
 	"scripts": {
-		"dev": "concurrently \"npm run dev --prefix backend/user_service\" \"npm run dev --prefix backend/game_service\" \"npm run dev --prefix frontend\" \"npm run dev --prefix backend/googleAuth\""
+		"dev": "concurrently \"npm run dev --prefix backend/user_service\" \"npm run dev --prefix backend/game_service\" \"npm run dev --prefix frontend\" \"npm run dev --prefix backend/auth_service\""
 	},
 	"devDependencies": {
 		"axios": "^1.8.4",
